updated - if re-install rename variables*.ps1

+++ functions/install_servervariables.ps1
@@ -9,9 +9,15 @@
 Function New-CreateVariables {
     Write-log "Function: New-CreateVariables"
     Write-log "Creating Variables Script"
-    New-Item $serverdir\Variables-$serverfiles.ps1 -Force | Out-File -Append -Encoding Default  $ssmlog
-#        Add-Content  $serverdir\Variables-$serverfiles.ps1  "#                            Discord Webhook "
-#        Add-Content  $serverdir\Variables-$serverfiles.ps1  "`$global:discordwebhook      = `"$global:discordwebhook`""
+    If (Test-Path $serverdir\Variables-$serverfiles.ps1) {
+        Rename-Item $serverdir\Variables-$serverfiles.ps1 $serverdir\Variables-$serverfiles.ps1.old
+        New-Item $serverdir\Variables-$serverfiles.ps1 -Force | Out-File -Append -Encoding Default  $ssmlog
+    }
+    Else {
+        New-Item $serverdir\Variables-$serverfiles.ps1 -Force | Out-File -Append -Encoding Default  $ssmlog
+    }
+    #        Add-Content  $serverdir\Variables-$serverfiles.ps1  "#                            Discord Webhook "
+    #        Add-Content  $serverdir\Variables-$serverfiles.ps1  "`$global:discordwebhook      = `"$global:discordwebhook`""
     If ($appid) {
         Add-Content  $serverdir\Variables-$serverfiles.ps1  "#                            App ID  "
         Add-Content  $serverdir\Variables-$serverfiles.ps1  "`$global:appid               = `"$global:appid`""
